[deps] Add and import Buffer, to avoid relying on a global definition. (#3)
This makes sense in a React Native app, where it's normal for Buffer
not to be defined. Note that in Node, this commit explicitly stops
using the Node builtin [1].

One can always define it globally themselves, as is done in
https://github.com/charpeni/react-native-url-polyfill/blob/master/index.js#L18.
(The version of 'buffer' added in this commit matches the 'buffer'
version in the latest master of that repository as of writing this.)
But it's nice, if feasible, to give the option to avoid changing
global variables.

The trailing slash was added as prompted in the Usage section of the
'buffer' page on NPM [2]; this means explicitly not using the Node
builtin [1].

To test that this works, I ran `npm run build` in the project
directory. Then, in my own React Native project's `package.json`, I
put '../whatwg-url' (the path to my local, built copy with this
change) for `whatwg-url-without-unicode` in my dependencies. My
project doesn't have anything assigned to the Buffer global
variable, nor does it directly depend on 'buffer' from NPM. It
worked as expected, without any `ReferenceError`s about Buffer not
being defined.

[1]: https://nodejs.org/api/url.html#url_class_url
[2]: https://www.npmjs.com/package/buffer#usage

Co-authored-by: Chris Bobbe <cbobbe@zulipchat.com>